refactor: 크루 저장 로직 분리 및 AuthorizationExtractor 패키지 이동
